+++ plan/projection_plan.hpp
@@ -33,8 +33,8 @@ class ProjectionPlan : public PlanBase {
       TransactionContext& ctx) const override;
   [[nodiscard]] Schema GetSchema(TransactionContext& ctx) const override;
 
-  [[nodiscard]] int AccessRowCount(TransactionContext& txn) const override;
-  [[nodiscard]] int EmitRowCount(TransactionContext& txn) const override;
+  [[nodiscard]] size_t AccessRowCount(TransactionContext& txn) const override;
+  [[nodiscard]] size_t EmitRowCount(TransactionContext& txn) const override;
   void Dump(std::ostream& o, int indent) const override;
 
  private:
